split FrontendSchemaBuilder
